fix Prod page
