--- src/pages/Login/components/Card/Card.css
@@ -1,13 +1,0 @@
-.Card {
-  background: #f5f5f5;
-  padding: 80px 105px;
-  padding-bottom: 205px;
-}
-
-.loginLogo {
-  margin-bottom: 85px;
-}
-
-.inputLogin {
-  margin-bottom: 46px;
-}
