+++ src/pages/ProductsPage/components/ProductsTable/ProductsTable.js
@@ -3,31 +3,23 @@ import { Component } from 'react';
 import { BiSortAlt2 } from "react-icons/bi";
 import { FaEdit } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
+import logo from "../../../../logo.svg";
 
 const sortIcon = <BiSortAlt2 className='sortIcon' size='20px'/>;
 const editIcon = <FaEdit className='editIcon' size='20px'/>;
 const deleteIcon = <MdDelete className='deleteIcon' size='20px' />
 
 class ProductsTable extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { products: [] }
-  }
-
-  componentDidMount(){
-    this.getProducts()
-  }
   
-  async getProducts() {
-    const response = await fetch('https://66401c9ca7500fcf1a9d1857.mockapi.io/products');
-    const productsData = await response.json();
-
-    this.setState({ products: productsData })
-    console.log(productsData);
-  }
-
   render(){
+
+    const { isLoading, isError, products } = this.props;
+
     return ( 
+
+      <div>
+
+      { isLoading ? (<img width="150px" height="150px" src={logo} className="App-logo" alt="logo" />) : (products.map((product) => (
       <div className='ProductsTableContainer'>
         <div className='productTableTitle'>
           <p className='slotName'>ID{sortIcon}</p>
@@ -39,18 +31,23 @@ class ProductsTable extends Component {
         </div>
 
         <div className='productTableField'>
-          <div className='productTableSelect'> { this.state.products.map(product => <div className='productItem'  key={product.id}> 
+        <div className='productTableSelect'>
+          <div className='productItem' key={product.id}> 
             <div className='productItem'>{product.id}</div> <div className='productItem'>{product.category}</div> 
               <div className='productItem' >{product.name}</div>  <div className='productItem' >{product.quantity}</div> 
               <div className='productItem' >{product.price}</div>
-                <div className='buttonContainerProductTable'>
-                  <button className='buttonEdit'>{editIcon}</button>
-                  <button className='buttonDelete'>{deleteIcon}</button></div> </div>)}
+                  <div className='buttonContainerProductTable'>
+                    <button className='buttonEdit'>{editIcon}</button>
+                    <button className='buttonDelete'>{deleteIcon}</button></div>
+                  </div>
               </div>
           </div>
+          {isError && <p>Oops! sorry we have a problem...........?</p> }
+          </div>
+          )))}
       </div>
-    );
+      )
+    }
   }
-}
 
 export default ProductsTable;
